requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.21
sqlalchemy>=1.4
psycopg2-binary>=2.9
alembic>=1.10
passlib[argon2]>=1.7
pyjwt>=2.8
python-dotenv>=1.0
pydantic-settings>=2.0
redis>=4.5
aioredis>=2.0
pytest>=7.3
pytest-asyncio>=0.21
httpx>=0.24
black>=24.3
ruff>=0.12
cryptography>=41.0
slowapi>=0.1.7

cryptography>=41.0
pyjwt>=2.8
redis>=4.5

